Moved IT's to use ReaktorRule and fixed compilation errors in tests.
