+++ src/test/java/org/reaktivity/nukleus/tcp/internal/streams/ClientLimitsIT.java
@@ -19,8 +19,6 @@ import static java.net.StandardSocketOptions.SO_REUSEADDR;
 import static java.util.concurrent.TimeUnit.SECONDS;
 import static org.junit.Assert.assertEquals;
 import static org.junit.rules.RuleChain.outerRule;
-import static org.reaktivity.nukleus.tcp.internal.TcpConfiguration.MAXIMUM_STREAMS_WITH_PENDING_WRITES;
-import static org.reaktivity.nukleus.tcp.internal.TcpConfiguration.WINDOW_SIZE;
 
 import java.io.IOException;
 import java.net.InetSocketAddress;
@@ -35,7 +33,8 @@ import org.junit.rules.TestRule;
 import org.junit.rules.Timeout;
 import org.kaazing.k3po.junit.annotation.Specification;
 import org.kaazing.k3po.junit.rules.K3poRule;
-import org.reaktivity.reaktor.test.NukleusRule;
+import org.reaktivity.reaktor.internal.ReaktorConfiguration;
+import org.reaktivity.reaktor.test.ReaktorRule;
 
 public class ClientLimitsIT
 {
@@ -45,19 +44,20 @@ public class ClientLimitsIT
 
     private final TestRule timeout = new DisableOnDebug(new Timeout(5, SECONDS));
 
-    private final NukleusRule nukleus = new NukleusRule("tcp")
+    private final ReaktorRule reaktor = new ReaktorRule()
+        .nukleus("tcp"::equals)
         .directory("target/nukleus-itests")
         .commandBufferCapacity(1024)
         .responseBufferCapacity(1024)
         .counterValuesBufferCapacity(1024)
-        .streams("tcp", "source#partition");
-
-    private final TestRule properties = new SystemPropertiesRule()
-        .setProperty(MAXIMUM_STREAMS_WITH_PENDING_WRITES.propertyName(), "1")
-        .setProperty(WINDOW_SIZE.propertyName(), "15");
+        // Initial window size for output to network:
+        .configure(ReaktorConfiguration.BUFFER_SLOT_CAPACITY_PROPERTY, 64)
+        // Overall buffer pool size same as slot size so maximum concurrent streams with partial writes = 1
+        .configure(ReaktorConfiguration.BUFFER_POOL_CAPACITY_PROPERTY, 64)
+        .clean();
 
     @Rule
-    public final TestRule chain = outerRule(properties).around(nukleus).around(k3po).around(timeout);
+    public final TestRule chain = outerRule(reaktor).around(k3po).around(timeout);
 
     @Test
     @Specification({
