+++ src/test/java/org/reaktivity/nukleus/tcp/internal/streams/ClientPartialWriteLimitsIT.java
@@ -24,8 +24,6 @@ import static java.util.stream.IntStream.of;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.rules.RuleChain.outerRule;
-import static org.reaktivity.nukleus.tcp.internal.TcpConfiguration.MAXIMUM_STREAMS_WITH_PENDING_WRITES;
-import static org.reaktivity.nukleus.tcp.internal.TcpConfiguration.WINDOW_SIZE;
 import static org.reaktivity.nukleus.tcp.internal.streams.SocketChannelHelper.ALL;
 
 import java.net.InetSocketAddress;
@@ -48,7 +46,8 @@ import org.kaazing.k3po.junit.rules.K3poRule;
 import org.reaktivity.nukleus.tcp.internal.TcpCountersRule;
 import org.reaktivity.nukleus.tcp.internal.streams.SocketChannelHelper.HandleWriteHelper;
 import org.reaktivity.nukleus.tcp.internal.streams.SocketChannelHelper.ProcessDataHelper;
-import org.reaktivity.reaktor.test.NukleusRule;
+import org.reaktivity.reaktor.internal.ReaktorConfiguration;
+import org.reaktivity.reaktor.test.ReaktorRule;
 
 /**
  * Tests the handling of capacity exceeded conditions in the context of incomplete writes
@@ -64,16 +63,17 @@ public class ClientPartialWriteLimitsIT
 
     private final TestRule timeout = new DisableOnDebug(new Timeout(5, SECONDS));
 
-    private final NukleusRule nukleus = new NukleusRule("tcp")
+    private final ReaktorRule reaktor = new ReaktorRule()
+        .nukleus("tcp"::equals)
         .directory("target/nukleus-itests")
         .commandBufferCapacity(1024)
         .responseBufferCapacity(1024)
         .counterValuesBufferCapacity(1024)
-        .streams("tcp", "source#partition");
-
-    private final TestRule properties = new SystemPropertiesRule()
-        .setProperty(MAXIMUM_STREAMS_WITH_PENDING_WRITES.propertyName(), "1")
-        .setProperty(WINDOW_SIZE.propertyName(), "15");
+        // Initial window size for output to network:
+        .configure(ReaktorConfiguration.BUFFER_SLOT_CAPACITY_PROPERTY, 64)
+        // Overall buffer pool size same as slot size so maximum concurrent streams with partial writes = 1
+        .configure(ReaktorConfiguration.BUFFER_POOL_CAPACITY_PROPERTY, 64)
+        .clean();
 
     private final TcpCountersRule counters = new TcpCountersRule()
         .directory("target/nukleus-itests")
@@ -82,8 +82,8 @@ public class ClientPartialWriteLimitsIT
         .counterValuesBufferCapacity(1024);
 
     @Rule
-    public final TestRule chain = outerRule(SocketChannelHelper.RULE).around(properties)
-                                  .around(nukleus).around(counters).around(k3po).around(timeout);
+    public final TestRule chain = outerRule(SocketChannelHelper.RULE)
+                                  .around(reaktor).around(counters).around(k3po).around(timeout);
 
     @Test
     @Specification({
